airflow/dags/utils/boe_scraper.py
"""
BOE scraper utility for fetching legislation documents.
"""
import requests
from datetime import datetime
from bs4 import BeautifulSoup
import os


BOE_API_BASE_URL = os.getenv('BOE_API_BASE_URL', 'https://www.boe.es')
BOE_API_TIMEOUT = int(os.getenv('BOE_API_TIMEOUT', '30'))


def fetch_boe_documents(start_date, end_date):
    """
    Fetch BOE documents between start_date and end_date.
    
    Note: This is a simplified implementation. In production, you would use
    the official BOE API or a more robust scraping solution.
    
    Args:
        start_date: datetime object for start date
        end_date: datetime object for end date
        
    Returns:
        List of document dictionaries
    """
    documents = []
    
    # This is a placeholder implementation
    # In a real scenario, you would:
    # 1. Query the BOE API: https://boe.es/datosabiertos/
    # 2. Parse the XML/JSON responses
    # 3. Extract document metadata and content
    
    # Example structure of what would be returned:
    example_doc = {
        'boe_id': 'BOE-A-2024-00001',
        'title': 'Example BOE Document',
        'summary': 'This is an example summary',
        'publication_date': datetime.now().date(),
        'document_type': 'Disposiciones generales',
        'department': 'Ministerio de Justicia',
        'section': 'I',
        'url': f'{BOE_API_BASE_URL}/boe/dias/2024/01/01/pdfs/BOE-A-2024-00001.pdf',
        'pdf_url': f'{BOE_API_BASE_URL}/boe/dias/2024/01/01/pdfs/BOE-A-2024-00001.pdf',
        'full_text': 'This is the full text content of the document...',
        'metadata': {
            'section': 'I',
            'subsection': 'Disposiciones generales',
            'organism': 'Ministerio de Justicia'
        }
    }
    
    # For demonstration purposes, return an empty list
    # In production, this would fetch real BOE documents
    print(f"Fetching BOE documents from {start_date} to {end_date}")
    print("Note: This is a placeholder. Implement real BOE API integration here.")
    
    return documents


def fetch_document_content(document_url):
    """
    Fetch the full text content of a BOE document.
    
    Args:
        document_url: URL of the document
        
    Returns:
        String containing the document text
    """
    try:
        response = requests.get(document_url, timeout=BOE_API_TIMEOUT)
        response.raise_for_status()
        
        # Parse HTML and extract text
        soup = BeautifulSoup(response.content, 'html.parser')
        
        # Remove script and style elements
        for script in soup(["script", "style"]):
            script.decompose()
        
        # Get text
        text = soup.get_text()
        
        # Clean up whitespace
        lines = (line.strip() for line in text.splitlines())
        chunks = (phrase.strip() for line in lines for phrase in line.split("  "))
        text = ' '.join(chunk for chunk in chunks if chunk)
        
        return text
    except Exception as e:
        print(f"Error fetching document content: {e}")
        return ""


def process_document(raw_document):
    """
    Process a raw BOE document into the standard format.
    
    Args:
        raw_document: Raw document data from BOE API
        
    Returns:
        Processed document dictionary
    """
    # This would process the raw XML/JSON from BOE API
    # into the standardized format used by the system
    return raw_document
