airflow/requirements.txt
psycopg2-binary==2.9.9
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.4
qdrant-client==1.7.0
langchain==0.1.0
langchain-community==0.0.10
